deepfake/pyproject.toml
[project]
name = "deepfake"
version = "0.1.0"
description = "Deepfake experiments with PyTorch Lightning"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "albumentations>=1.4",
    "click>=8.0",
    "lightning>=2.0",
    "omegaconf>=2.3",
    "pillow>=10.0",
    "torch",
]

[project.scripts]
deepfake = "deepfake.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deepfake"]
